requirements.txt
streamlit
ultralytics
opencv-python
Pillow
numpy
torch
torchvision
scikit-learn